Stopgap fix for locally numbered preallocation - This works, but overestimates the sparsity structure - Refined point renumbering
Hg-commit: 86b97699220ab139f160b0a8ed819d1dc16b2f3e

+++ src/dm/mesh/examples/tests/idistribution1.cxx
@@ -199,7 +199,7 @@ public:
     sieve->base(baseV);
     const sieve_type::point_type *base = baseV.getPoints();
 
-    for(int b = 0, k = 0; b < baseV.getSize(); ++b) {
+    for(int b = 0, k = 0; b < (int) baseV.getSize(); ++b) {
       CPPUNIT_ASSERT_EQUAL(this->_renumbering[coneRoots[b]], base[b]);
       sieve->cone(base[b], cV);
       const sieve_type::point_type *cone     = cV.getPoints();
@@ -232,7 +232,7 @@ public:
     sieve->cap(capV);
     const sieve_type::point_type *cap = capV.getPoints();
 
-    for(int c = 0, k = 0; c < capV.getSize(); ++c) {
+    for(int c = 0, k = 0; c < (int) capV.getSize(); ++c) {
       CPPUNIT_ASSERT_EQUAL(this->_renumbering[supportRoots[c]], cap[c]);
       sieve->support(cap[c], sV);
       const sieve_type::point_type *support     = sV.getPoints();
@@ -319,7 +319,7 @@ public:
     size_t localSize;
     f >> localSize;
     int *ordering = new int[localSize*3];
-    for(int i = 0; i < localSize; ++i) {
+    for(int i = 0; i < (int) localSize; ++i) {
       f >> ordering[i*3+0]; // point
       f >> ordering[i*3+1]; // offset
       f >> ordering[i*3+2]; // dim
